+++ bin/install.sh
@@ -296,7 +296,6 @@ function _main ()
     local SHOW_HELP=0
 
     local PATH_TO_CONFIGURATION_FILE="${CURRENT_SCRIPT_PATH}/../data/configuration.sh"
-    local PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY=""   #will be filled up later
 
     while true;
     do
@@ -352,23 +351,6 @@ function _main ()
         . "${PATH_TO_CONFIGURATION_FILE}"
     fi
     
-    if [[ -f /usr/bin/pacman ]];
-    then
-        _echo_if_be_verbose ":: Packagemanager >>pacman<< detected."
-        PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY="/etc/zabbix/zabbix_agentd.conf.d"
-
-    elif [[ -f /usr/bin/apt ]];
-    then
-        _echo_if_be_verbose ":: Packagemanager >>apt<< detected."
-        PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY="/etc/zabbix/zabbix_agentd.d"
-
-    else
-        _echo_if_be_verbose ":: No supported package manager found."
-        _echo_if_be_verbose "   pacman or apt are mandatory right now. Feel free to create a pull request to support more package managers."
-
-        exit 1
-    fi
-
     if [[ ${IS_DRY_RUN} -eq 1 ]];
     then
         local TEMPORARY_DIRECTORY=$(mktemp -d)
@@ -385,15 +367,15 @@ function _main ()
         local FILE_PATH_TO_SYSTEMD_SERVICE_FILE="${TEMPORARY_DIRECTORY}${FILE_PATH_TO_SYSTEMD_SERVICE_FILE}"
         local FILE_PATH_TO_SYSTEMD_TIMER_FILE="${TEMPORARY_DIRECTORY}${FILE_PATH_TO_SYSTEMD_TIMER_FILE}"
         local FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT="${TEMPORARY_DIRECTORY}${FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT}"
-        local PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY="${TEMPORARY_DIRECTORY}${PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY}"
+        local DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION="${TEMPORARY_DIRECTORY}${DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION}"
 
         local DIRECTORY_PATH_FOR_PACKAGES_FILES=$(dirname "${FILE_PATH_TO_REGULAR_PACKAGES}")
 
         echo "   Creating path >>${DIRECTORY_PATH_FOR_PACKAGES_FILES}<<."
         mkdir -p "${DIRECTORY_PATH_FOR_PACKAGES_FILES}"
 
-        echo "   Creating path >>${PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY}<<."
-        mkdir -p "${PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY}"
+        echo "   Creating path >>${DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION}<<."
+        mkdir -p "${DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION}"
     fi
 
     if [[ ${BE_VERBOSE} -eq 1 ]];
@@ -414,7 +396,7 @@ function _main ()
         echo "   FILE_PATH_TO_SYSTEMD_SERVICE_FILE >>${FILE_PATH_TO_SYSTEMD_SERVICE_FILE}<<"
         echo "   FILE_PATH_TO_SYSTEMD_TIMER_FILE >>${FILE_PATH_TO_SYSTEMD_TIMER_FILE}<<"
         echo "   FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT >>${FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT}<<"
-        echo "   PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY >>${PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY}<<"
+        echo "   DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION >>${DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION}<<"
         echo ""
     fi
 
@@ -426,6 +408,11 @@ function _main ()
     elif [[ -f /usr/bin/apt ]];
     then
         _create_script_for_apt "${FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT}" "${FILE_PATH_TO_REGULAR_PACKAGES}" "${FILE_PATH_TO_SECURITY_PACKAGES}"
+    else
+        _echo_if_be_verbose ":: No supported package manager found."
+        _echo_if_be_verbose "   pacman or apt are mandatory right now. Feel free to create a pull request to support more package managers."
+
+        exit 2
     fi
 
     chmod +x "${FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT}"
@@ -433,7 +420,7 @@ function _main ()
     #take a look on zabbix_mysql_housekeeping/bin/install.sh
     _create_systemd_files "${FILE_PATH_TO_PACKAGE_FILES_GENERATION_SCRIPT}" "${FILE_PATH_TO_SYSTEMD_SERVICE_FILE}" "${FILE_PATH_TO_SYSTEMD_TIMER_FILE}"
 
-    _add_zabbix_agent_configuration "${PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY}" "${FILE_PATH_TO_REGULAR_PACKAGES}" "${FILE_PATH_TO_SECURITY_PACKAGES}" "${ZABBIX_AGENT_CONFIGURATION_NAME}"
+    _add_zabbix_agent_configuration "${DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION}" "${FILE_PATH_TO_REGULAR_PACKAGES}" "${FILE_PATH_TO_SECURITY_PACKAGES}" "${ZABBIX_AGENT_CONFIGURATION_NAME}"
 
     _echo_if_be_verbose ":: Finished installation"
     _echo_if_be_verbose "   Please import the template file in path >>${CURRENT_SCRIPT_PATH}/../template/update_notifyer.xml<< in your zabbix server."