+++ CHANGELOG.md
@@ -12,12 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * add check for `cat /etc/zabbix/zabbix_agentd.conf | grep -v '^#\|^$' | grep '^Include='`
     * if grep does not get a hit, we need to add the path to then end of the configuration file
 * add `${ZABBIX_CONFIGURATION_FILE_NAME}`
-    * `local ZABBIX_CONFIGURATION_FILE_NAME="update_notifyer.conf`
 * add `README.md#Idea`
 
 ### To Change
-
-* change zabbix configuration directory to `/usr/local/etc/zabbix_agentd.conf.d`
 
 ## [Unreleased]
 
@@ -28,7 +25,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+* Change zabbix configuration directory to `/usr/local/etc/zabbix_agentd.conf.d`
 * Instead of stopping the script, we are now starting the same script with sudo prefix again
+* Replaced >>PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION_DIRECTORY<< with >>DIRECTORY_PATH_TO_THE_ZABBIX_AGENT_CONFIGURATION<<
 * Replaced >>ROOT_PATH_TO_PACKAGE_CONFIGURATION<< with >>DIRECTORY_PATH_TO_PACKAGE_CONFIGURATION<<
 
 ## [0.1.0](https://github.com/bazzline/zabbix_agent_update_notifyer/tree/0.1.0) - released at 20220317